Rename to SpendAmount to be more general & move
