+++ cli/src/spend_utils.rs
@@ -1,13 +1,37 @@
 use crate::{
     checks::{calculate_fee, check_account_for_balance},
-    cli::{CliError, TransferAmount},
+    cli::CliError,
 };
+use clap::ArgMatches;
+use solana_clap_utils::{input_parsers::lamports_of_sol, offline::SIGN_ONLY_ARG};
 use solana_client::rpc_client::RpcClient;
 use solana_sdk::{
     fee_calculator::FeeCalculator, message::Message, native_token::lamports_to_sol, pubkey::Pubkey,
     transaction::Transaction,
 };
 use std::error;
+
+#[derive(Debug, PartialEq, Clone, Copy)]
+pub enum SpendAmount {
+    All,
+    Some(u64),
+}
+
+impl SpendAmount {
+    pub fn new(amount: Option<u64>, sign_only: bool) -> Self {
+        match amount {
+            Some(lamports) => Self::Some(lamports),
+            None if !sign_only => Self::All,
+            _ => panic!("ALL amount not supported for sign-only operations"),
+        }
+    }
+
+    pub fn new_from_matches(matches: &ArgMatches<'_>, name: &str) -> Self {
+        let amount = lamports_of_sol(matches, name);
+        let sign_only = matches.is_present(SIGN_ONLY_ARG.name);
+        SpendAmount::new(amount, sign_only)
+    }
+}
 
 struct SpendAndFee {
     spend: u64,
@@ -17,7 +41,7 @@ struct SpendAndFee {
 pub fn resolve_spend_tx_and_check_account_balance<F, G>(
     rpc_client: &RpcClient,
     sign_only: bool,
-    amount: TransferAmount,
+    amount: SpendAmount,
     fee_calculator: &FeeCalculator,
     from_pubkey: &Pubkey,
     build_message: F,
@@ -42,7 +66,7 @@ where
 pub fn resolve_spend_tx_and_check_account_balances<F, G>(
     rpc_client: &RpcClient,
     sign_only: bool,
-    amount: TransferAmount,
+    amount: SpendAmount,
     fee_calculator: &FeeCalculator,
     from_pubkey: &Pubkey,
     fee_pubkey: &Pubkey,
@@ -98,7 +122,7 @@ where
 }
 
 fn resolve_spend_message<F>(
-    amount: TransferAmount,
+    amount: SpendAmount,
     fee_calculator: &FeeCalculator,
     from_balance: u64,
     from_pubkey: &Pubkey,
@@ -109,7 +133,7 @@ where
     F: Fn(u64) -> Message,
 {
     match amount {
-        TransferAmount::Some(lamports) => {
+        SpendAmount::Some(lamports) => {
             let message = build_message(lamports);
             let fee = calculate_fee(fee_calculator, &[&message]);
             (
@@ -120,7 +144,7 @@ where
                 },
             )
         }
-        TransferAmount::All => {
+        SpendAmount::All => {
             let dummy_message = build_message(0);
             let fee = calculate_fee(fee_calculator, &[&dummy_message]);
             let lamports = if from_pubkey == fee_pubkey {
