+++ cli/tests/nonce.rs
@@ -1,12 +1,13 @@
 use solana_cli::test_utils::check_balance;
 use solana_cli::{
-    cli::{process_command, request_and_confirm_airdrop, CliCommand, CliConfig, TransferAmount},
+    cli::{process_command, request_and_confirm_airdrop, CliCommand, CliConfig},
     cli_output::OutputFormat,
     nonce,
     offline::{
         blockhash_query::{self, BlockhashQuery},
         parse_sign_only_reply_string,
     },
+    spend_utils::SpendAmount,
 };
 use solana_client::rpc_client::RpcClient;
 use solana_core::contact_info::ContactInfo;
@@ -311,7 +312,7 @@ fn test_create_account_with_seed() {
     authority_config.command = CliCommand::ClusterVersion;
     process_command(&authority_config).unwrap_err();
     authority_config.command = CliCommand::Transfer {
-        amount: TransferAmount::Some(10),
+        amount: SpendAmount::Some(10),
         to: to_address,
         from: 0,
         sign_only: true,
@@ -333,7 +334,7 @@ fn test_create_account_with_seed() {
         format!("http://{}:{}", leader_data.rpc.ip(), leader_data.rpc.port());
     submit_config.signers = vec![&authority_presigner];
     submit_config.command = CliCommand::Transfer {
-        amount: TransferAmount::Some(10),
+        amount: SpendAmount::Some(10),
         to: to_address,
         from: 0,
         sign_only: false,
